initial test reorganization
