+++ hadoop-tools/hadoop-posum/src/test/java/org/apache/hadoop/tools/posum/database/monitor/TestHadoopMonitor.java
@@ -1,6 +1,7 @@
-package org.apache.hadoop.tools.posum.test;
+package org.apache.hadoop.tools.posum.database.monitor;
 
 import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.tools.posum.common.util.POSUMConfiguration;
 import org.apache.hadoop.tools.posum.database.master.DataMasterContext;
 import org.apache.hadoop.tools.posum.database.client.DBInterface;
 import org.apache.hadoop.tools.posum.database.store.DataStore;
@@ -12,11 +13,11 @@ import javax.xml.crypto.Data;
 /**
  * Created by ane on 3/3/16.
  */
-public class TestSystemMonitor {
+public class TestHadoopMonitor {
 
     @Test
     public void checkDatabaseFeeding() {
-        Configuration conf = TestUtils.getConf();
+        Configuration conf = POSUMConfiguration.newInstance();
         DataStore dataStore = new DataStore(conf);
         DataMasterContext context = new DataMasterContext();
         context.setDataStore(dataStore);
