+++ hadoop-tools/hadoop-posum/src/test/java/org/apache/hadoop/tools/posum/test/CustomTests.java
@@ -6,7 +6,7 @@ import org.junit.Test;
 /**
  * Created by ane on 4/20/16.
  */
-public class TestConnectivity {
+public class CustomTests {
     @Test
     public void checkRegistration(){
         System.out.println(NetUtils.createSocketAddr("0.0.0.0", 7000));
