--- hadoop-tools/hadoop-posum/src/test/java/org/apache/hadoop/tools/posum/test/TestUtils.java
@@ -1,14 +1,0 @@
-package org.apache.hadoop.tools.posum.test;
-
-import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.tools.posum.common.util.POSUMConfiguration;
-
-/**
- * Created by ane on 3/3/16.
- */
-public class TestUtils {
-
-    static Configuration getConf() {
-        return POSUMConfiguration.newInstance();
-    }
-}
